back/requirements.txt
anyio==3.3.2
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
fastapi==0.70.0
h11==0.12.0
httptools==0.2.0
idna==3.2
pycparser==2.20
pydantic==1.8.2
PyJWT==2.2.0
PyMySQL==1.0.2
python-dotenv==0.19.0
PyYAML==5.4.1
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing-extensions==3.10.0.2
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
